Allow spaces in the header list for proper formatting
The datatable auto-aligns it's column width based on the word in the
header. If that word was too short the values could wrap and look
strange. Now we allow spaces in the header list to widen the column a
bit, but ignore the space when using the header value as a key.